apps/strands-agent/pyproject.toml
[project]
name = "strands-venice-agent"
version = "0.1.0"
description = "Venice GLM-4.7 uncensored research agent — Strands Agents SDK"
requires-python = ">=3.10"
dependencies = [
    "strands-agents",
    "strands-agents-tools",
    "strands-agents-builder",
    "openai",
    "python-dotenv",
    "fastapi[standard]",
    "uvicorn[standard]",
    "ddgs",
    "httpx",
    "pdfplumber",
    "yt-dlp",
    "EbookLib",
    "beautifulsoup4",
    "lxml",
    "b2sdk",
    "rocrate",
    "duckdb",
    "langchain-openai",
    "deepagents>=0.5.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "agent", "config", "tools", "prompts", "cache", "youtube_tools", "document_tools", "book_pipeline", "datalake", "extraction", "mcp_configs", "jobs", "swarm_bridge", "corpus", "atomizer", "corpus_tools", "orchestrator", "orchestrator_protocol", "orchestrator_langchain", "task_pool", "task_tools", "knowledge_store", "knowledge_recall_tools"]
